Pix2Vox/requirements.txt
argparse
easydict
matplotlib
numpy
scipy
torchvision
tensorboardX
